move better auth instance to lib folder
